This is an example of Vector
In C++, a vector is a dynamic array that can grow or shrink in size. It is part of the Standard Template Library (STL) and is defined in the <vector> header. Vectors provide a flexible and efficient way to store and manipulate collections of elements. Here's a brief description of vectors in C++: